refactor: Code cleanup
